adding split to images and fixing bug in recolor

+++ recolor/recolor.py
@@ -42,12 +42,12 @@ class ColorSpace:
     def get_o2(self):
 
         o2 = np.array( map(self.bgr2o2, np.reshape(self.image,(self.width*self.height,self.chans))),dtype=np.float32)
-        return np.reshape(o2,(self.height,self.width))
+        return np.reshape(o2,(self.height,self.width)).astype(np.uint8,copy=False)
         
     def get_o3(self):
         
         o3 = np.array(map(self.bgr2o3,np.reshape(self.image,(self.width*self.height,self.chans))),dtype=np.float32)
-        return np.reshape(o3,(self.height,self.width))
+        return np.reshape(o3,(self.height,self.width)).astype(np.uint8,copy=False)
         
     def bgr2o2(self,x):
